lib/models/requirements.txt
langchain
langchain-community
langchain-openai
openai
pdfplumber
pymupdf
python-dotenv
pytesseract
pdf2image
Pillow